역할 분리 및 리펙토링
