+++ src/main/kotlin/racingcar/Driver.kt
@@ -1,11 +1,11 @@
 package racingcar
 
 class Driver(private val car: Car) {
+    val driveRecord = mutableListOf<Int>()
     fun drive(isMovable: Boolean) {
-        if (isMovable) car.move()
-    }
-
-    fun getMoveResult(): String {
-        return "-".repeat(car.movement)
+        if (isMovable) {
+            car.move()
+        }
+        driveRecord.add(car.movement)
     }
 }
